Show data contained in ExecutedActionMetadata
This change lets bb_browser display all of the information that is
stored in the ExecutedActionMetadata message. This means that the name
of the worker and timing information is now shown in the page.

It also adds logic for showing messages that are specific to Buildbarn,
namely POSIX resource usage and FUSE file system statistics.