bb_browser: Don't show a scatter plot when no data is available
If you have the ISCC enabled for platforms that only have a single size
class, it is likely that you will end up with ISCC entries that have no
samples in them, but only have last_seen_failure set. This means that a
scatter plot consisting of zero columns would need to be drawn, causing
a crash on the following line:

    p.NominalX(sizeClassLabels...)

Address this by omitting the scatter plot if no samples are present.