Let bb_browser display file system access profiles
bb_worker is now capable of recording which files in input roots are
actually being accessed during execution, and storing that in a new data
stored named the File System Access Cache (FSAC) in the form of a Bloom
filter.

This change extends bb_browser to display this information on action
pages. This is useful when designing build rules, and want to know which
files can be removed from the input root.

https://github.com/buildbarn/bb-adrs/blob/master/0010-file-system-access-cache.md