Unbreak PurgeCSS rules that use multiple files
When switching to rules_js, we had to prefix ${PWD} to all pathnames, as
rules_js ends up changing the working directory internally. This
silently broke the places where we used this in combination with
$(locations ...), as that may yield a sequence of multiple pathnames. We
need to prepend ${PWD} to each of the paths. PurgeCSS doesn't warn if
files cannot be found.